backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djoser==2.2.2
psycopg2-binary==2.9.9
Pillow==10.3.0
django-cleanup==8.1.0
django-filter==2.4.0
gunicorn==21.2.0
fpdf2==2.7.8
uharfbuzz==0.39.1
drf_extra_fields==3.7.0